pyproject.toml
[project]
name = "ambigen"
version = "0.1.0"
description = "Generate ambient YouTube videos from YAML recipes"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "click>=8.0",
    "google-generativeai>=0.8",
    "pyyaml>=6.0",
    "pillow>=10.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[project.scripts]
ambigen = "ambigen.cli:cli"

[tool.setuptools.packages.find]
include = ["ambigen*"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
